requirements.txt
langchain==0.1.0
langchain-community==0.0.10
langchain-groq==0.0.1
chromadb==0.4.22
pypdf2==3.0.1
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0
numpy>=2.0.0
pandas>=2.2.0
fastembed==0.3.2
networkx==3.2.1
plotly==5.17.0
requests==2.31.0
Pillow==10.1.0
setuptools>=68.0.0
wheel>=0.40.0
